+++ docs/changes.rst
@@ -3,6 +3,17 @@
 Change Log
 ==========
 
+0.1.3 - May 6, 2015
+-------------------
+
+* Add LabelValueReadView
+
+0.1.2 - May 6, 2015
+-------------------
+
+* Use six for python 2/3 compatibility
+* expose model verbose names to templates
+
 0.1.1 - April 30, 2015
 ----------------------